+++ src/test/java/uk/gov/hmcts/reform/idam/service/IdamDuplicateUserMergerServiceTest.java
@@ -41,7 +41,7 @@ class IdamDuplicateUserMergerServiceTest {
     IdamClient idamClient;
 
     @Mock
-    UserRoleService userRoleService;
+    UserRoleMergeService userRoleMergeService;
 
     @Mock
     LoggingSummaryUtils summaryUtils;
